+++ frontend/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next'
 import './globals.css'
 
 // Components
-import Navbar from '../components/Navbar'
+import Navbar from '@/components/Navbar'
 
 export const metadata: Metadata = {
   title: 'Deepfake Detector',
