+++ frontend/app/loading.tsx
@@ -1,0 +1,23 @@
+export default function Loading() {
+	return (
+		<main className="z-50 bg-whitehue-600 w-screen h-screen flex items-center justify-center relative opacity-70">
+			<div className="flex flex-col items-center">
+				<div className="h-20 w-40 flex items-center">
+					<span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-300"></span>
+					<span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-100"></span>
+					<span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-0"></span>
+					<span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-700"></span>
+					<span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-150"></span>
+					<span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-500"></span>
+					<span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-75"></span>
+					<span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-100"></span>
+					<span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-500"></span>
+					<span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-300"></span>
+					<span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-150"></span>
+					<span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-0"></span>
+				</div>
+				<h1 className="mt-3">Loading...</h1>
+			</div>
+		</main>
+	)
+}
