--- frontend/app/loading/page.tsx
@@ -1,22 +1,0 @@
-export default function Loading() {
-  return (
-    <div className="flex flex-col items-center">
-      <div className="h-20 w-40 flex items-center">
-        <span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-300"></span>
-        <span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-100"></span>
-        <span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-0"></span>
-        <span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-700"></span>
-        <span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-150"></span>
-        <span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-500"></span>
-        <span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-75"></span>
-        <span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-100"></span>
-        <span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-500"></span>
-        <span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-300"></span>
-        <span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-150"></span>
-        <span className="bg-black block h-full w-3 rounded-3xl mx-1 animate-soundwave animate-delay-0"></span>
-      </div>
-      <h1 className="mt-3">Loading...</h1>
-    </div>
-
-  )
-}
